разбиение на слои
