+++ home_work2/sum.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+isValidNumber() {
+    if [[ "$1" =~ ^[+-]?([0-9]+\.?[0-9]*|\.[0-9]+)$ ]]; then
+        return 0
+    else
+        return 1
+    fi
+}
