+++ home_work4/php/app.php
@@ -1,0 +1,9 @@
+<?php
+require 'route.php';
+class App {
+
+    public function run()
+    {
+      Route::route();
+    }
+}
